@@ -14,17 +14,17 @@ classifiers = [
   "License :: OSI Approved :: MIT License",
   "Programming Language :: Python",
   "Programming Language :: Python :: 3",
-  "Programming Language :: Python :: 3.8",
   "Programming Language :: Python :: 3.9",
   "Programming Language :: Python :: 3.10",
   "Programming Language :: Python :: 3.11",
+  "Programming Language :: Python :: 3.12",
   "Topic :: Internet :: WWW/HTTP",
   "Framework :: AsyncIO",
   "Framework :: Trio",
 ]
 
 [tool.poetry.dependencies]
-python = "^3.8"
+python = "^3.9"
 apispec = "^6"
 marshmallow = "^3"
 muffin = "^0"
@@ -67,7 +67,7 @@ marshmallow-peewee = "^4.0.3"
 
 
 [tool.pytest.ini_options]
-addopts = "-xsv"
+addopts = "-xsv tests"
 log_cli = true
 
 [tool.mypy]
@@ -77,7 +77,7 @@ ignore_missing_imports = true
 [tool.tox]
 legacy_tox_ini = """
 [tox]
-envlist = py38,py39,py310,py311,pypy39
+envlist = py39,py310,py311,py312,pypy39
 
 [testenv]
 deps = -e .[tests]
@@ -87,7 +87,7 @@ commands =
 
 [tool.ruff]
 line-length = 100
-target-version = "py38"
+target-version = "py39"
 exclude = [".venv", "docs", "examples"]
 select = ["ALL"]
 ignore = [
@@ -112,7 +112,7 @@ ignore = [
 
 [tool.black]
 line-length = 100
-target-version = ["py38", "py39", "py310", "py311"]
+target-version = ["py39", "py310", "py311", "py312"]
 preview = true
 
 [build-system]
