@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Dict
+from typing import TYPE_CHECKING
 
 from muffin_rest.openapi import OpenAPIMixin
 
@@ -19,7 +19,7 @@ class PeeweeOpenAPIMixin(OpenAPIMixin):
     meta: PWRESTOptions
 
     @classmethod
-    def openapi(cls, route: Route, spec: APISpec, tags: Dict) -> Dict:
+    def openapi(cls, route: Route, spec: APISpec, tags: dict) -> dict:
         """Get openapi specs for the endpoint."""
         operations = super(PeeweeOpenAPIMixin, cls).openapi(route, spec, tags)
         is_resource_route = getattr(route, "params", {}).get(cls.meta.name_id)
